MAINT: refactor all math to global utils
